+++ core/hyphenator-liang.lua
@@ -114,7 +114,7 @@ local hyphenateNode = function (node)
     local hyphen = SILE.shaper:createNnodes(SILE.settings:get("font.hyphenchar"), node.options)
     local newnodes = {}
     for j, segment in ipairs(segments) do
-      local leadingApostrophe
+      local specificDiscretionary
       if segment == "" then
         SU.dump({ j, segments })
         SU.error("No hyphenation segment should ever be empty", true)
@@ -124,8 +124,40 @@ local hyphenateNode = function (node)
         if nextApostrophe then
           segments[j+1] = luautf8.gsub(segments[j+1], "^['’]", "")
           local replacement = SILE.shaper:createNnodes(nextApostrophe, node.options)
-          leadingApostrophe = SILE.nodefactory.discretionary({ replacement = replacement, prebreak = hyphen })
-          leadingApostrophe.parent = node
+          if SILE.settings:get("languages.tr.replaceApostropheAtHyphenation") then
+            -- leading apostrophe (on next segment) cancels when hyphenated
+            specificDiscretionary = SILE.nodefactory.discretionary({ replacement = replacement, prebreak = hyphen })
+          else
+            -- hyphen character substituted for upcomming apostrophe
+            local kesme = SILE.shaper:createNnodes(nextApostrophe, node.options)
+            specificDiscretionary = SILE.nodefactory.discretionary({ replacement = replacement, prebreak = kesme })
+          end
+        end
+      elseif node.options.language == "ca" then
+        -- punt volat (middle dot) cancels when hyphenated
+        -- Catalan typists may use a punt volat or precomposed characters.
+        -- The shaper might behave differently depending on the font, so we need to
+        -- be consistent here with the typist's choice.
+        if luautf8.find(segment, "ŀ$") then -- U+0140
+          segment = luautf8.sub(segment, 1, -2)
+          local ldot = SILE.shaper:createNnodes("ŀ", node.options)
+          local lhyp = SILE.shaper:createNnodes("l" .. SILE.settings:get("font.hyphenchar"), node.options)
+          specificDiscretionary = SILE.nodefactory.discretionary({ replacement = ldot, prebreak = lhyp })
+        elseif luautf8.find(segment, "Ŀ$") then -- U+013F
+          segment = luautf8.sub(segment, 1, -2)
+          local ldot = SILE.shaper:createNnodes("Ŀ", node.options)
+          local lhyp = SILE.shaper:createNnodes("L" .. SILE.settings:get("font.hyphenchar"), node.options)
+          specificDiscretionary = SILE.nodefactory.discretionary({ replacement = ldot, prebreak = lhyp })
+        elseif luautf8.find(segment, "l·$") then -- l + U+00B7
+          segment = luautf8.sub(segment, 1, -3)
+          local ldot = SILE.shaper:createNnodes("l·", node.options)
+          local lhyp = SILE.shaper:createNnodes("l" .. SILE.settings:get("font.hyphenchar"), node.options)
+          specificDiscretionary = SILE.nodefactory.discretionary({ replacement = ldot, prebreak = lhyp })
+        elseif luautf8.find(segment, "L·$") then -- L + U+00B7
+          segment = luautf8.sub(segment, 1, -3)
+          local ldot = SILE.shaper:createNnodes("L·", node.options)
+          local lhyp = SILE.shaper:createNnodes("L" .. SILE.settings:get("font.hyphenchar"), node.options)
+          specificDiscretionary = SILE.nodefactory.discretionary({ replacement = ldot, prebreak = lhyp })
         end
       end
       for _, newNode in ipairs(SILE.shaper:createNnodes(segment, node.options)) do
@@ -135,8 +167,9 @@ local hyphenateNode = function (node)
         end
       end
       if j < #segments then
-        if leadingApostrophe then
-          table.insert(newnodes, leadingApostrophe)
+        if specificDiscretionary then
+          specificDiscretionary.parent = node
+          table.insert(newnodes, specificDiscretionary)
         else
           local newNode = SILE.nodefactory.discretionary({ prebreak = hyphen })
           newNode.parent = node